services/auth-service/requirements.txt
django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
redis==5.0.8
django-cors-headers==4.4.0
gunicorn==22.0.0
pydantic==2.8.2
python-decouple==3.8
